FEAT: Update bridge UI to aggregator style UI (#410)
* make bridge ui into aggregator

* no assertion

* fix edge cases

* mobile layout

* redundant return

* adjust origin to destination wormhole mapping

* apply review

* fix another init state function of useState

* optimise BigNumber usage
