--- src/components/bridge/BridgeFallback.tsx
@@ -1,9 +1,0 @@
-import { LoadingBar } from "../common/LoadingBar";
-
-export const BridgeFallback = () => {
-  return (
-    <div className="flex justify-center">
-      <LoadingBar />
-    </div>
-  );
-};
