+++ src/utils/helpers/mutationCallback.ts
@@ -1,10 +1,20 @@
-import { wagmiConfig } from "@/lib/wagmi/wagmiConfig";
+import { UseMutationOptions } from "@tanstack/react-query";
+import {
+  SendTransactionErrorType,
+  SendTransactionParameters,
+  SendTransactionReturnType,
+  WriteContractErrorType,
+  WriteContractParameters,
+  WriteContractReturnType,
+} from "@wagmi/core";
 import { toast } from "sonner";
 import {
   TransactionNotFoundError,
   WaitForTransactionReceiptTimeoutError,
 } from "viem";
-import { UsePublicClientReturnType, UseWriteContractParameters } from "wagmi";
+import { UsePublicClientReturnType } from "wagmi";
+
+import { wagmiConfig } from "@/lib/wagmi/wagmiConfig";
 
 interface MutationCallbackArgs {
   action: string;
@@ -19,7 +29,11 @@ export const mutationCallback = ({
   action,
   addRecentTransaction,
   publicClient,
-}: MutationCallbackArgs): UseWriteContractParameters["mutation"] =>
+}: MutationCallbackArgs): UseMutationOptions<
+  WriteContractReturnType | SendTransactionReturnType,
+  WriteContractErrorType | SendTransactionErrorType,
+  WriteContractParameters | SendTransactionParameters
+> =>
   ({
     onSuccess: (hash) => {
       addRecentTransaction({
